Automatic detection of metadata fields
Have added the programmatic detection of metadata fields to the 'get_dataset_subelements()'
function, and therefore to all subsequent functions that call it.  This allows 'delete_project()'
to identify the data fields that are present automatically.  This utilizes the 'get_entity_info()'
function.

As part of this effort, created the 'search_entity()' function to automatically
use the correct 'search_*' function for a given entity.  Function is based on
the workings of the 'get_entity' function.  This function should probably
be moved to the "functions.R" file.

Have also added the ability to ignore error messages that arrise from searching in fields
that do not have defined search functions, despite indications that they should (this
currently occurs for only the 'EXPERIMENT' entity, which is in the middle of being
implemented.